spectral profile region, stats and stokes dropdowns moved to toolbar
