+++ src/components/SpectralProfiler/SpectralProfileToolbarComponent/SpectralProfileToolbarComponent.scss
@@ -1,0 +1,13 @@
+.spectral-profiler-toolbar {
+    overflow: hidden;
+
+    .bp3-form-group {
+        margin-bottom: 5px;
+        margin-top: 5px;
+        margin-left: 15px;
+
+        &:first-of-type {
+            margin-left: 5px;
+        }
+    }
+}
