+++ src/components/SpectralProfiler/SpectralProfilerComponent.scss
@@ -12,8 +12,9 @@
         flex: 3 3 auto;
         min-width: 0;
         padding-bottom: 5px;
+
         .profile-plot {
-            height: 100%;
+            height: calc(100% - 35px);
             position: relative;
         }
     }
